requirements.txt
attrs==19.3.0
bitmex==0.2.2
bokeh==2.0.2
bravado==10.6.0
bravado-core==5.17.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
MarkupSafe==1.1.1
monotonic==1.5
msgpack==1.0.0
msgpack-python==0.5.6
numpy==1.18.2
packaging==20.4
pandas==1.0.3
Pillow==7.1.2
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
rfc3987==1.3.8
simplejson==3.17.0
six==1.14.0
SQLAlchemy==1.3.17
strict-rfc3339==0.7
swagger-spec-validator==2.5.0
TA-Lib==0.4.17
tornado==6.0.4
typing-extensions==3.7.4.2
urllib3==1.25.9
webcolors==1.11.1
Werkzeug==1.0.1
WTForms==2.3.1
